Batching: adds support for maxBatchSize option
As discussed in #39 there are certain situations where a user
may want to limit the amount of requests sent to the batchUrl at once.

This PR changes the batched request workflow so that if maxBatchSize is
provided it will chunk out the batched requests into numerous requests,
the flatten the responses into a single array and resolve them the same
way relay-network-layer already does.

Some utility functions were added to help stay DRY and limit the amount
of looping necessary when no maxBatchSize is provided.